+++ ghedt/examples/Design/BiRectangular/find_bi_rectangle.py
@@ -149,7 +149,7 @@ def main():
     fig, ax = ghedt.gfunction.GFunction.visualize_area_and_constraints(
         perimeter, coordinates, no_go=no_go)
 
-    fig.savefig('bi-rectangle_case.png')
+    fig.savefig('bi-rectangle_case.png', bbox_inches='tight', pad_inches=0.1)
 
 
 if __name__ == '__main__':