+++ ghedt/examples/Design/BiZonedRectangle/find_bi_zoned_rectangle.py
@@ -1,5 +1,5 @@
 # Jack C. Cook
-# Thursday, October 28, 2021
+# Saturday, October 30, 2021
 
 import ghedt
 import ghedt.PLAT as PLAT
@@ -95,20 +95,36 @@ def main():
     # Rectangular design constraints are the land and range of B-spacing
     length = 85.  # m
     width = 36.5  # m
-    B_min = 3.  # m
-    B_max = 10.  # m
+    B_min = 4.45  # m
+    B_max_x = 10.  # m
+    B_max_y = 12.
+
+    _coordinates_domain_nested = \
+        ghedt.domains.bi_rectangle_zoned_nested(
+            length, width, B_min, B_max_x, B_max_y)
+
+    output_folder = 'Bi-Rectangle_Zoned_Domain_Original/'
+    # for i in range(len(_coordinates_domain_nested)):
+    #     coordinates_domain = _coordinates_domain_nested[i]
+    #     ghedt.domains.visualize_domain(coordinates_domain,
+    #                                    output_folder + str(i))
 
     # Perform field selection using bisection search between a 1x1 and 32x32
-    coordinates_domain = ghedt.domains.rectangular(length, width, B_min, B_max)
+    # coordinates_domain_nested = \
+    #     ghedt.domains.bi_zoned_domain_restructured(
+    #         length, width, B_min, B_max_x, B_max_y)
 
-    output_folder = 'Rectangle_Domain'
-    ghedt.domains.visualize_domain(coordinates_domain, output_folder)
+    output_folder = 'Bi-Rectangle_Zoned_Domain_Restructured/'
+    # for i in range(len(coordinates_domain_nested)):
+    #     coordinates_domain = coordinates_domain_nested[i]
+    #     ghedt.domains.visualize_domain(coordinates_domain,
+    #                                    output_folder + str(i))
 
     tic = clock()
-    bisection_search = ghedt.search_routines.Bisection1D(
-        coordinates_domain, V_flow_borehole, borehole, bhe_object,
+    bisection_search = ghedt.search_routines.BisectionZD(
+        _coordinates_domain_nested, V_flow_borehole, borehole, bhe_object,
         fluid, pipe, grout, soil, sim_params, hourly_extraction_ground_loads,
-        disp=False)
+        disp=True)
     toc = clock()
     print('Time to perform bisection search: {} seconds'.format(toc - tic))
 
@@ -145,7 +161,8 @@ def main():
     fig, ax = ghedt.gfunction.GFunction.visualize_area_and_constraints(
         perimeter, coordinates, no_go=no_go)
 
-    fig.savefig('base_case.png', bbox_inches='tight', pad_inches=0.1)
+    fig.savefig('bi-rectangle_zoned_case.png', bbox_inches='tight',
+                pad_inches=0.1)
 
 
 if __name__ == '__main__':